wip: migrate microlearning pages to new path structure
