+++ apps/frontend-manage/src/components/activities/creation/microLearning/MicroLearningWizard.tsx
@@ -329,7 +329,7 @@ function MicroLearningWizard({
           )}
           name={formData.name}
           editMode={editMode}
-          previewElementHref={`${process.env.NEXT_PUBLIC_PWA_URL}/course/${selectedCourseId}/microlearning/${microLearningCreateData?.createMicroLearning?.id || microLearningEditData?.editMicroLearning?.id}/`}
+          previewElementHref={`${process.env.NEXT_PUBLIC_PWA_URL}/course/${selectedCourseId}/microLearnings/${microLearningCreateData?.createMicroLearning?.id || microLearningEditData?.editMicroLearning?.id}/`}
           viewElementHref={`/courses/${selectedCourseId}?tab=microLearnings`}
           onRestartForm={() => {
             setIsWizardCompleted(false)
