+++ apps/frontend-manage/src/components/courses/MicroLearningList.tsx
@@ -21,14 +21,14 @@ function MicroLearningList({
         <QRCodePopover
           triggerStyle="basic"
           triggerText={t('manage.general.qrCode')}
-          relHref={`/course/${courseId}/microLearnings`}
+          relHref={`/course/${courseId}/microLearnings/overview`}
           data={{ cy: `qr-link-practice-quiz-list` }}
         />
         <Button
           basic
           onClick={async () => {
             try {
-              const link = `${process.env.NEXT_PUBLIC_LTI_URL}?redirectTo=${process.env.NEXT_PUBLIC_PWA_URL}/course/${courseId}/microLearnings`
+              const link = `${process.env.NEXT_PUBLIC_LTI_URL}?redirectTo=${process.env.NEXT_PUBLIC_PWA_URL}/course/${courseId}/microLearnings/overview`
               console.log(link)
               await navigator.clipboard.writeText(link)
             } catch (e) {
